requirements.txt
imageio==2.9.0
matplotlib==3.3.4
networkx==2.4
numpy==1.18.5
open3d==0.10.0.0
opencv-contrib-python==4.3.0.36
opencv-python==4.3.0.36
Pillow==7.2.0
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
termcolor==1.1.0
torch==1.5.1
torchvision==0.6.1
torchviz==0.0.1
tqdm==4.47.0
transforms3d==0.3.1
bm3d==3.0.7